+++ app/models/provider.rb
@@ -1,8 +1,4 @@
 class Provider < ApplicationRecord
-  # TODO: temporary measure until we integrate properly
-  # with LAA Portal (current SAML mock lacks email)
-  FALLBACK_EMAIL = 'provider@example.com'.freeze
-
   devise :lockable, :timeoutable, :trackable,
          :omniauthable, omniauth_providers: %i[saml]
 
@@ -24,12 +20,10 @@ class Provider < ApplicationRecord
     def from_omniauth(auth)
       find_or_initialize_by(auth_provider: auth.provider, uid: auth.uid).tap do |record|
         record.update(
-          email: auth.info.email.presence || FALLBACK_EMAIL,
+          email: auth.info.email,
           description: auth.info.description,
-          roles: auth.info.roles.split(','),
-          # SAMLmock returns comma-separated values, however other
-          # Portal envs (dev, staging, production) use colons
-          office_codes: auth.info.office_codes.split(/[:,]/),
+          roles: auth.info.roles,
+          office_codes: auth.info.office_codes,
         )
 
         ensure_default_office(record)
