+++ spec/requests/errors_spec.rb
@@ -38,7 +38,7 @@ RSpec.describe 'Error pages' do
     context 'when the sign in fails' do
       before do
         allow(OmniAuth.config).to receive(:test_mode).and_return(false)
-        allow_any_instance_of(LaaPortalSetup).to receive(:setup).and_raise(StandardError)
+        allow_any_instance_of(LaaPortal::SamlSetup).to receive(:setup).and_raise(StandardError)
       end
 
       it 'redirects to the home' do
